Faz ajustes em menu
