+++ proc/wagtail_hooks.py
@@ -13,11 +13,18 @@ from wagtail.contrib.modeladmin.views import CreateView, InspectView
 
 from config.menu import get_menu_order
 from package.models import SPSPkg
+from htmlxml.models import HTMLXML
 
 from .models import ArticleProc, IssueProc, JournalProc
 
 
 class ProcCreateView(CreateView):
+    def form_valid(self, form):
+        self.object = form.save_all(self.request.user)
+        return HttpResponseRedirect(self.get_success_url())
+
+
+class CoreCreateView(CreateView):
     def form_valid(self, form):
         self.object = form.save_all(self.request.user)
         return HttpResponseRedirect(self.get_success_url())
@@ -92,6 +99,48 @@ class IssueProcModelAdmin(ModelAdmin):
     )
 
 
+class HTMLXMLModelAdmin(ModelAdmin):
+    model = HTMLXML
+    menu_label = _("XML from HTML")
+    menu_icon = "doc-full"
+    menu_order = 300
+    add_to_settings_menu = False
+    exclude_from_explorer = True
+    inspect_view_enabled = True
+
+    list_per_page = 10
+    create_view_class = CoreCreateView
+
+    list_display = (
+        "article_proc",
+        "html2xml_status",
+        "quality",
+        "attention_demands",
+        "html_translation_langs",
+        "pdf_langs",
+        "n_paragraphs",
+        "n_references",
+        "created_updated",
+    )
+    list_filter = (
+        "html_img_total",
+        "html_table_total",
+        "empty_body",
+        "attention_demands",
+        "article_type",
+        "html2xml_status",
+        "quality",
+        "html_translation_langs",
+        "pdf_langs",
+    )
+    search_fields = (
+        "article_proc__migrated_data__pid",
+        "article_proc__pkg_name",
+        "html2xml_status",
+        "article_type",
+    )
+
+
 class SPSPkgModelAdmin(ModelAdmin):
     model = SPSPkg
     menu_label = _("SPS Package")
@@ -104,6 +153,10 @@ class SPSPkgModelAdmin(ModelAdmin):
     list_display = (
         "pid_v3",
         "sps_pkg_name",
+        "is_pid_provider_synchronized",
+        "valid_texts",
+        "valid_components",
+        "is_public",
         "xml_uri",
         "created",
         "updated",
@@ -166,8 +219,9 @@ class ProcessModelAdminGroup(ModelAdminGroup):
     items = (
         JournalProcModelAdmin,
         IssueProcModelAdmin,
-        ArticleProcModelAdmin,
+        HTMLXMLModelAdmin,
         SPSPkgModelAdmin,
+        ArticleProcModelAdmin,
     )
